cambie el tipo de dato de creditos_aprob_total

+++ backEnd/gestion-asignacion/gestion_hojas_de_calculo/models.py
@@ -20,7 +20,7 @@ class PerfilAcademico(models.Model):
     # Campos que se llenarán con el Excel (mapeo directo)
     promedio = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     num_electivas_cursadas = models.IntegerField(default=0)
-    creditos_aprob_total = models.DecimalField(max_digits=10, decimal_places=2, default=0)
+    creditos_aprob_total =  models.IntegerField(default=0)
     num_periodos_matriculados = models.IntegerField(default=0)
     
     # Campos con valores por defecto/calculados/fijos (se definen en el Resource)
